# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for generating scenarios and scenes using Artificial Intelligence.",
        "title": "Scene AI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@scene-ai.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/ai/scenario": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Trigger AI to generate a new scenario based on a prompt.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Generate AI scenario",
                "parameters": [
                    {
                        "description": "Generation prompt",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.GenerateScenarioRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Job accepted, returns scenario ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/scenario/{id}": {
            "get": {
                "description": "Retrieve the status and details of an AI-generated scenario.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Get AI scenario status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scenario ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Scenario details",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.Scenario"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/scenario/{id}/scenes": {
            "get": {
                "description": "Retrieve the list of generated scenes for a scenario.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Get generated scenes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scenario ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of scenes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.Scene"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/scenes/{id}": {
            "get": {
                "description": "Establish a WebSocket connection to stream scene generation progress.",
                "tags": [
                    "AI"
                ],
                "summary": "Connect to scene stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scene ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticate user and return access token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.Token"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Create a new user account with email and password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.User"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    },
                    "409": {
                        "description": "User already exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve user profile by ID or current token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get user details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (optional, defaults to token owner)",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.User"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently delete the current user's account.",
                "tags": [
                    "Auth"
                ],
                "summary": "Delete user account",
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update current user's information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "Update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated user",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.User"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            }
        },
        "/api/v1/scenario/scenario": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a manual scenario with scenes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scenario"
                ],
                "summary": "Create a new scenario",
                "parameters": [
                    {
                        "description": "Scenario details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.CreateScenarioRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Scenario created",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.Scenario"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            }
        },
...
        "/api/v1/scenario/scenarios": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of scenarios with pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scenario"
                ],
                "summary": "List scenarios",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of the last item from previous page",
                        "name": "before-id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of scenarios",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.Scenario"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.ErrMsg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_devathh_scene-ai_internal_common_dtos.ErrMsg": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_devathh_scene-ai_internal_common_dtos.GenerateScenarioRequest": {
            "type": "object",
            "properties": {
                "prompt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_devathh_scene-ai_internal_common_dtos.Scenario": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "global_style_prompt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "scenario_prompt": {
                    "type": "string"
                },
                "scenes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_devathh_scene-ai_internal_common_dtos.Scene"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}